Grafic output of my t-SNE code
I use t-SNE method to compare the impact from three of tested parameters for diabetes development. It allows to see which two of them taken together affects diabetes the most. We can see by the distribution of points in different colors that parameters 0 and 1 together allow to see the biggest difference in the result. 

I use both, PCA and t-SNE method to compare, which of them is a better tool in our case. We can see that results on both methods lead to similar conclusions, but PCA is more transparent in our case.

In my work I compared only three of ten parameters. The next step, which should be taken is running PCA and t-SNE for other parameters and choose the most important ones.